change WidthMillis back to int

+++ api/base_test.go
@@ -13,7 +13,7 @@ import (
 
 func TestDoMove(t *testing.T) {
 	dev := &inject.Base{}
-	dev.WidthMillisFunc = func(ctx context.Context) (float64, error) {
+	dev.WidthMillisFunc = func(ctx context.Context) (int, error) {
 		return 600, nil
 	}
 	ang, dist, err := api.DoMove(context.Background(), api.Move{}, dev)
