+++ testutils/inject/base.go
@@ -11,7 +11,7 @@ type Base struct {
 	api.Base
 	MoveStraightFunc func(ctx context.Context, distanceMillis int, millisPerSec float64, block bool) (int, error)
 	SpinFunc         func(ctx context.Context, angleDeg float64, degsPerSec float64, block bool) (float64, error)
-	WidthMillisFunc  func(ctx context.Context) (float64, error)
+	WidthMillisFunc  func(ctx context.Context) (int, error)
 	StopFunc         func(ctx context.Context) error
 	CloseFunc        func() error
 }
@@ -30,7 +30,7 @@ func (b *Base) Spin(ctx context.Context, angleDeg float64, degsPerSec float64, b
 	return b.SpinFunc(ctx, angleDeg, degsPerSec, block)
 }
 
-func (b *Base) WidthMillis(ctx context.Context) (float64, error) {
+func (b *Base) WidthMillis(ctx context.Context) (int, error) {
 	if b.WidthMillisFunc == nil {
 		return b.Base.WidthMillis(ctx)
 	}
